Open `AvatarBitmapFactory` to allow custom implementation
