+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/common/internal/ImageLoaderFactoryInitializer.kt
@@ -8,7 +8,7 @@ import coil.decode.GifDecoder
 import coil.decode.ImageDecoderDecoder
 import com.getstream.sdk.chat.coil.StreamCoil
 import com.getstream.sdk.chat.coil.StreamImageLoaderFactory
-import io.getstream.chat.android.ui.avatar.internal.AvatarBitmapFactory
+import io.getstream.chat.android.ui.avatar.AvatarBitmapFactory
 
 @Suppress("unused")
 internal class ImageLoaderFactoryInitializer : Initializer<ImageLoaderFactory> {
